fix: unify model icon resolution
